# TaskiePet.Client/src/api/generated/models/api-response-of-object.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TaskiePet.WebApi | v1
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface ApiResponseOfObject {
    'isSuccess'?: boolean;
    'message'?: string | null;
    'data'?: any;
}