# TaskiePet.Client/src/api/generated/models/api-response-of-token-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TaskiePet.WebApi | v1
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { TokenResponse } from './token-response';

export interface ApiResponseOfTokenResponse {
    'isSuccess'?: boolean;
    'message'?: string | null;
    'data'?: TokenResponse | null;
}