# TaskiePet.Client/src/api/generated/models/daily-task.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TaskiePet.WebApi | v1
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface DailyTask {
    'userId'?: string;
    'title'?: string;
    'description'?: string;
    'isCompleted'?: boolean;
    'completedAt'?: string | null;
    'id'?: string;
    'createdAt'?: string;
    'updatedAt'?: string;
}